refactor: add render function for running without the CLI
