+++ src/cli/types.ts
@@ -1,0 +1,15 @@
+export type RenderCliOptions = {
+  // Directory to save output files
+  outDir?: string
+  // Run browser in headless mode
+  headless?: boolean
+  // Watch resume file for changes
+  watch?: boolean
+  // Theme name to use
+  theme?: string
+}
+
+export type InitCliOptions = {
+  // Theme name to use
+  theme?: string
+}
