+++ src/types.ts
@@ -8,20 +8,6 @@ export interface Resume {
   }
 }
 
-export type RenderOptions = {
-  // Directory to save output files
-  outDir?: string
-  // Run browser in headless mode
-  headless?: boolean
-  // Watch resume file for changes
-  watch?: boolean
-  // Theme name to use
-  theme?: string
-}
-
-export type InitOptions = {
-  // Theme name to use
-  theme?: string
-}
+export type RenderOptions = { theme: string; outDir: string }
 
 export type ConsoleLog = typeof console.log
